create gap cursor when clicking at very top of body
