+++ src/gwt/panmirror/src/editor/src/nodes/div.ts
@@ -14,13 +14,12 @@
  */
 
 import { Node as ProsemirrorNode, Schema, DOMOutputSpec, ResolvedPos } from 'prosemirror-model';
-import { EditorState, Transaction, PluginKey, Plugin } from 'prosemirror-state';
+import { EditorState, Transaction } from 'prosemirror-state';
 import { EditorView } from 'prosemirror-view';
 import { findParentNodeOfType, ContentNodeWithPos, findParentNodeOfTypeClosestToPos } from 'prosemirror-utils';
 import { wrapIn } from 'prosemirror-commands';
 import { GapCursor } from 'prosemirror-gapcursor';
 import { liftTarget } from 'prosemirror-transform';
-import { keymap } from 'prosemirror-keymap';
 
 import { ExtensionContext, Extension } from '../api/extension';
 import {
@@ -122,24 +121,9 @@ const extension = (context: ExtensionContext) : Extension | null => {
     baseKeys: () => {
       return [
         { key: BaseKey.Enter, command: divInputRuleEnter() },
+        { key: BaseKey.ArrowLeft, command: arrowHandler('left') },
+        { key: BaseKey.ArrowUp, command: arrowHandler('up') }
       ];
-    },
-
-    plugins: (schema: Schema) => {
-      return [
-        new Plugin({
-          key: new PluginKey('div-gap-cursor'),
-          props: {
-            handleDOMEvents: {
-              click: clickHandler,
-            },
-          },
-        }),
-        keymap({
-          ArrowLeft: arrowHandler('left'),
-          ArrowUp: arrowHandler('up')
-        }),
-      ];  
     },
 
     commands: () => {
@@ -160,39 +144,6 @@ const extension = (context: ExtensionContext) : Extension | null => {
   };
 };
 
-function clickHandler(view: EditorView, event: Event): boolean {
-
-  const mouseEvent = event as MouseEvent;
-  const clickPos = view.posAtCoords({ left: mouseEvent.clientX, top: mouseEvent.clientY } );
-
-  if (clickPos) {
-
-    const div = findParentNodeOfTypeClosestToPos(
-      view.state.doc.resolve(clickPos.pos), view.state.schema.nodes.div
-    );
-    if (div && div.pos === clickPos.inside) {
-      
-      // focus the view
-      view.focus();
-      
-      // create the gap cursor
-      const tr = view.state.tr;
-      const $gapPos = tr.doc.resolve(clickPos.pos);
-      const gapCursor = new GapCursor($gapPos, $gapPos); 
-      tr.setSelection(gapCursor);
-      view.dispatch(tr);
-      
-      // prevent default event handling
-      event.preventDefault();
-      event.stopImmediatePropagation();
-      return true;
-    }
-    
-
-  }
-
-  return false;
-}
 
 function arrowHandler(_dir: 'up' | 'left') {
   return (state: EditorState, dispatch?: (tr: Transaction) => void, view?: EditorView) => {
